Split out uploadTheme
